Fix for gcc.

+++ DataReader/UCIFastReader/UCIFastReader.cpp
@@ -385,7 +385,7 @@ void UCIFastReader<ElemType>::InitFromConfig(const ConfigRecordType & readerConf
         fprintf(stderr, "Reading UCI file %ls\n", file.c_str());
 
     // Simple heuristic to ensure buffer size and avoid breaking existing experiments.
-    size_t bufSize = max(dimFeatures * 16, 256 * 1024);
+    size_t bufSize = max(dimFeatures * 16, (size_t)256 * 1024);
     m_parser.ParseInit(file.c_str(), startFeatures, dimFeatures, startLabels, dimLabels, bufSize);
 
     // if we have labels, we need a label Mapping file, it will be a file with one label per line
